implement database crud operations

+++ lib/core/db/impl/sqlite_database.dart
@@ -7,6 +7,8 @@ import 'package:inventory_management_with_sql/core/db/utils/sql_utils.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:sqflite/sqflite.dart';
 
+// SqliteDatabase? _single;
+
 class SqliteDatabase implements DataStore<Database> {
   final String dbName;
   final int version;
@@ -15,7 +17,25 @@ class SqliteDatabase implements DataStore<Database> {
   SqliteDatabase._(this.dbName, [this.version = 1]);
 
   ///singletone
+  ///dbName => SqliteDatabase
   static final Map<String, SqliteDatabase> _instance = {};
+
+  // static SqliteDatabase? _single;
+
+  // static SqliteDatabase get single {
+  //   _single ??= SqliteDatabase._('hello');
+  //   return _single!;
+  // }
+
+  // static SqliteDatabase single() {
+  //   _single ??= SqliteDatabase._("dbName");
+  //   return _single!;
+  // }
+
+  // factory SqliteDatabase.single() {
+  //   _single ??= SqliteDatabase._("dbName");
+  //   return _single!;
+  // }
 
   static SqliteDatabase newInstance(String dbName, [int version = 1]) {
     _instance[dbName] ??= SqliteDatabase._(dbName, version);
@@ -39,11 +59,17 @@ class SqliteDatabase implements DataStore<Database> {
     database = await openDatabase(
       dbFile.path,
       version: version,
-      onCreate: (db, __) async {
-        await onUp(db);
+      onCreate: (db, current) async {
+        await onUp(
+          current,
+          db,
+        );
       },
-      onDowngrade: (db, __, ___) async {
-        await onDown(db);
+      onUpgrade: (db, old, current) async {
+        await onUp(current, db);
+      },
+      onDowngrade: (db, old, current) async {
+        await onDown(old, current, db);
       },
     );
   }
@@ -56,22 +82,43 @@ class SqliteDatabase implements DataStore<Database> {
   }
 
   @override
-  Future<void> onUp([Database? db]) async {
-    if (db == null) {
-      assert(database != null);
-    }
+  Future<void> onUp(
+    int version, [
+    Database? db,
+  ]) async {
+    assert(db != null || database != null);
 
-    await Future.wait(tableNames.map((tableName) {
+    final migration = tableNames[version];
+    final columnMigration = tableColumns[version];
+    if (migration == null || columnMigration == null) throw "version not found";
+
+    await Future.wait(migration.map((tableName) {
       String query = """Create table if not exists "$tableName" (
         id integer primary key autoincrement,
         created_at text not null,  
         updated_at text,
       """;
 
-      for (TableProperties column in tableColumns[tableName] ?? []) {
+      ///other column
+      for (TableProperties column in columnMigration[tableName] ?? []) {
         query += toSqlQuery(column);
       }
+
+      /// Create table if not exists $tableName (
+      ///  id integer primary key autoincrement,
+      ///  created_at text not null,
+      ///  updated_at text,
+      ///  name varchar(255) not null,
+      ///  length > index  > 1
+      ///  2
       query = query.replaceFirst(",", "", query.length - 2);
+
+      /// Create table if not exists $tableName (
+      ///  id integer primary key autoincrement,
+      ///  created_at text not null,
+      ///  updated_at text,
+      ///  name varchar(255) not null
+
       query += ");";
       return (db ?? database)!.execute(query);
     }));
@@ -79,15 +126,22 @@ class SqliteDatabase implements DataStore<Database> {
   }
 
   @override
-  Future<void> onDown([Database? db]) async {
-    if (db == null) {
-      assert(database != null);
+  Future<void> onDown(int old, current, [Database? db]) async {
+    assert(db != null || database != null);
+    final currentMigration = tableNames[current];
+    final oldMigration = tableNames[old];
+    final oldColumnMigration = tableColumns[old];
+    if (currentMigration == null ||
+        oldMigration == null ||
+        oldColumnMigration == null) {
+      throw "version not found";
     }
-    await Future.wait(tableNames.reversed.map((e) {
+    await Future.wait(currentMigration.reversed.map((e) {
       return (db ?? database)!.execute(""" 
-        drop table if exists "$e" 
+        drop table if exists "$e"; 
       """);
     }));
     print("object deleted");
+    await onUp(old, db);
   }
 }
